+++ CHANGES.rst
@@ -11,6 +11,10 @@ Changelog
 2.1.0 (2014-10-23)
 ------------------
 
+- Read the enable_link_integrity_checks setting from plone.app.registry
+  instead of from the portal_properties.
+  [timo]
+
 - Restructure package to fully support dexterity framework. Use two different
   test layers in ``plone.app.testing``, migrate all doctests into real
   TestCases.